Move set_mock_attr into core next to the other mock helpers
set_mock_attr belongs with set_mock_value / callback_on_mock_put in
ophyd_async.core rather than in ophyd_async.testing. It is new in this PR, so
no back-compat shim is needed.
