+++ src/ophyd_async/testing/__init__.py
@@ -24,7 +24,6 @@ from ._assert import (
     assert_value,
     partial_reading,
 )
-from ._mock_attr import set_mock_attr
 from ._one_of_everything import (
     ExampleEnum,
     ExampleSubsetEnum,
@@ -109,5 +108,4 @@ __all__ = [
     "MovableBeamstop",
     "ReadOnlyBeamstop",
     "assert_has_calls",
-    "set_mock_attr",
 ]
